Handle object store tags
